# openapi/nhms.v1.yaml
openapi: 3.1.0
info:
  title: NHMS API
  description: National Hydrological Modeling System API
  version: 0.1.0
paths:
...
  /api/v1/basin-versions/{basin_version_id}/river-segments/{segment_id}/forecast-series:
    get:
      tags:
      - forecast
      summary: Get Forecast Series
      operationId: get_forecast_series_api_v1_basin_versions__basin_version_id__river_segments__segment_id__forecast_series_get
      parameters:
      - name: basin_version_id
        in: path
        required: true
        schema:
          type: string
          title: Basin Version Id
      - name: segment_id
        in: path
        required: true
        schema:
          type: string
          title: Segment Id
      - name: river_network_version_id
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          description: River network version for the selected/list row; required because
            river_segment_id is only unique within a river network version.
          title: River Network Version Id
        description: River network version for the selected/list row; required because
          river_segment_id is only unique within a river network version.
      - name: issue_time
        in: query
        required: false
        schema:
          type: string
          default: latest
          title: Issue Time
      - name: variables
        in: query
        required: false
        schema:
          type: string
          default: q_down
          title: Variables
      - name: scenarios
        in: query
        required: false
        schema:
          type: string
          default: GFS
          title: Scenarios
      - name: include_analysis
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Analysis
      - name: run_types
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Run Types
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Forecast Series Api V1 Basin Versions  Basin Version
                  Id  River Segments  Segment Id  Forecast Series Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}:
    get:
      tags:
      - forecast
      summary: Get Run
      operationId: get_run_api_v1_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Run Api V1 Runs  Run Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs:
    get:
      tags:
      - forecast
      summary: List Runs
      operationId: list_runs_api_v1_runs_get
      parameters:
      - name: basin_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Basin Id
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
      - name: cycle_time
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cycle Time
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Runs Api V1 Runs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/mvp/qhh/latest-product:
    get:
      tags:
      - runs
      summary: Get latest QHH display product
      operationId: getQhhLatestProduct
      parameters:
      - name: source
        in: query
        required: true
        schema:
          type: string
          enum:
          - GFS
          - IFS
        description: MVP forecast source. Accepted case-insensitively and normalized
          to GFS or IFS.
      - name: basin_id
        in: query
        required: false
        schema:
          type: string
        description: Target basin id for the latest display product. Defaults to basins_qhh
          when omitted, preserving backward compatibility for /api/v1/mvp/qhh/latest-product
          and M22 cross-plane callers.
      - name: run_id
        in: query
        required: false
        schema:
          type: string
          minLength: 1
        description: Strict QHH run identity. If supplied, source, cycle_time, and
          model_id must also be supplied; the API will not fall back to source-only
          latest selection.
      - name: cycle_time
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: Strict QHH cycle time. If supplied, source, run_id, and model_id
          must also be supplied; the API will not fall back to source-only latest
          selection.
      - name: model_id
        in: query
        required: false
        schema:
          type: string
          minLength: 1
        description: Strict QHH model identity. If supplied, source, run_id, and cycle_time
          must also be supplied; the API will not fall back to source-only latest
          selection.
      - name: identity_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
        description: 'Lightweight resolver for popups: returns run identity + cycle
          + horizon (+ recent cycles as available_issue_times) WITHOUT the expensive
          station/segment coverage computation. cycle_time may be supplied alone to
          select a specific cycle.'
      responses:
        '200':
          description: Latest QHH display product
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessEnvelope'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/QhhLatestProduct'
        4XX:
          $ref: '#/components/responses/Error'
        5XX:
          $ref: '#/components/responses/Error'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    SuccessEnvelope:
      type: object
      description: Standard success envelope returned by API endpoints using the `_ok()`
        response pattern.
      required:
      - request_id
      - status
      properties:
        request_id:
          type: string
          example: req_01J0NHMS
        status:
          type: string
          enum:
          - ok
          example: ok
    ErrorResponse:
      type: object
      required:
      - request_id
      - status
      - error
      properties:
        request_id:
          type: string
          example: req_01J0NHMS
        status:
          type: string
          enum:
          - error
          example: error
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Requested resource was not found.
            details:
              oneOf:
              - type: object
                nullable: true
                additionalProperties: true
              - type: array
                items:
                  $ref: '#/components/schemas/ValidationErrorDetail'
    ValidationErrorDetail:
      type: object
      required:
      - field
      - reason
      properties:
        field:
          type: string
        rejected_value:
          oneOf:
          - type: string
            nullable: true
          - type: number
          - type: boolean
          - type: object
            additionalProperties: true
          - type: array
            items: {}
        reason:
          type: string
      additionalProperties: true
...
    QhhLatestUnavailableReason:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
        run_id:
          type: string
          nullable: true
        source_id:
          type: string
          nullable: true
      additionalProperties: true
    QhhLatestQualityNote:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
        expected_horizon_hours:
          type: integer
          nullable: true
        available_horizon_hours:
          type: integer
          nullable: true
        available_end_time:
          type: string
          format: date-time
          nullable: true
      additionalProperties: true
    QhhLatestStationVariableCoverage:
      type: object
      required:
      - variable
      - station_count
      - sample_count
      - unit_count
      - quality_flag_count
      - missing_unit_samples
      - missing_quality_flag_samples
      - valid_time_start
      - valid_time_end
      properties:
        variable:
          type: string
          enum:
          - PRCP
          - TEMP
          - RH
          - wind
          - Rn
          - Press
        station_count:
          type: integer
          minimum: 0
        sample_count:
          type: integer
          minimum: 0
        unit_count:
          type: integer
          minimum: 0
        quality_flag_count:
          type: integer
          minimum: 0
        missing_unit_samples:
          type: integer
          minimum: 0
        missing_quality_flag_samples:
          type: integer
          minimum: 0
        valid_time_start:
          type: string
          format: date-time
          nullable: true
        valid_time_end:
          type: string
          format: date-time
          nullable: true
    QhhLatestQueryIndex:
      type: object
      required:
      - table
      - index
      - status
      - columns
      properties:
        table:
          type: string
        index:
          type: string
        status:
          type: string
        columns:
          type: array
          items:
            type: string
        predicate:
          type: string
          nullable: true
      additionalProperties: true
    QhhLatestAvailability:
      type: object
      required:
      - ready
      - unavailable_reasons
      - quality_flags
      - quality_notes
      properties:
        ready:
          type: boolean
        unavailable_reasons:
          type: array
          items:
            $ref: '#/components/schemas/QhhLatestUnavailableReason'
        quality_flags:
          type: array
          items:
            type: string
        quality_notes:
          type: array
          items:
            $ref: '#/components/schemas/QhhLatestQualityNote'
    QhhLatestQuality:
      type: object
      required:
      - station_sample_count
      - river_sample_count
      - required_station_variables
      - station_variable_coverage
      - candidate_limit
      - search_limit
      - context_limit
      - query_indexes
      properties:
        station_sample_count:
          type: integer
          minimum: 0
        river_sample_count:
          type: integer
          minimum: 0
        required_station_variables:
          type: array
          items:
            type: string
            enum:
            - PRCP
            - TEMP
            - RH
            - wind
            - Rn
            - Press
        station_variable_coverage:
          type: array
          items:
            $ref: '#/components/schemas/QhhLatestStationVariableCoverage'
        candidate_limit:
          type: integer
          minimum: 1
        search_limit:
          type: integer
          minimum: 1
        context_limit:
          type: integer
          minimum: 1
        query_indexes:
          type: array
          items:
            $ref: '#/components/schemas/QhhLatestQueryIndex'
    QhhLatestProduct:
      type: object
      required:
      - basin_id
      - model_id
      - basin_version_id
      - river_network_version_id
      - source_id
      - cycle_time
      - run_id
      - forcing_version_id
      - station_count
      - expected_station_count
      - segment_count
      - expected_segment_count
      - status
      - run_status
      - valid_time_start
      - valid_time_end
      - river_valid_time_start
      - river_valid_time_end
      - forcing_valid_time_start
      - forcing_valid_time_end
      - available_horizon_hours
      - expected_horizon_hours
      - shorter_horizon
      - availability
      - quality
      properties:
        basin_id:
          type: string
        model_id:
          type: string
        basin_version_id:
          type: string
        river_network_version_id:
          type: string
        available_issue_times:
          type: array
          items:
            type: string
            format: date-time
          description: Recent forecast cycles (newest first) for the issue-time selector.
            Only populated by the identity_only resolver.
        source_id:
          type: string
          enum:
          - GFS
          - IFS
        cycle_time:
          type: string
          format: date-time
        run_id:
          type: string
        forcing_version_id:
          type: string
        station_count:
          type: integer
          minimum: 0
        expected_station_count:
          type: integer
          minimum: 0
          nullable: true
        segment_count:
          type: integer
          minimum: 0
        expected_segment_count:
          type: integer
          minimum: 0
          nullable: true
        status:
          type: string
          enum:
          - ready
          - unavailable
        run_status:
          type: string
        valid_time_start:
          type: string
          format: date-time
          nullable: true
        valid_time_end:
          type: string
          format: date-time
          nullable: true
        river_valid_time_start:
          type: string
          format: date-time
          nullable: true
        river_valid_time_end:
          type: string
          format: date-time
          nullable: true
        forcing_valid_time_start:
          type: string
          format: date-time
          nullable: true
        forcing_valid_time_end:
          type: string
          format: date-time
          nullable: true
        available_horizon_hours:
          type: integer
          nullable: true
        expected_horizon_hours:
          type: integer
        shorter_horizon:
          type: boolean
        availability:
          $ref: '#/components/schemas/QhhLatestAvailability'
        quality:
          $ref: '#/components/schemas/QhhLatestQuality'
...
  responses:
...
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...